image를 만든다.
이미지를 먼저 만들고 저장한다.

dress를 만든다. dress를 만들고 저장한다. 만드는 과정에서 에러가 발생하면? image를 지우고 파일저장을 지운다.

+++ shopping_back/src/main/java/shopping/back/hj/dress/DressRepository.java
@@ -4,7 +4,6 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DressRepository extends JpaRepository<Dress, Integer> {
-
+public interface DressRepository extends JpaRepository<Dress, Long> {
 	Optional<Dress> findById(Long id);
 }
