+++ shopping_back/src/test/java/shopping/back/hj/dress/DressTest.java
@@ -15,7 +15,7 @@ public class DressTest {
 	
 	@Test
 	@Parameters(method = "paramsForTest")
-	public void builder(String brand, String article_num, Sex sex, Integer price, DressType dresstype, Integer discount) {
+	public void builder(String brand, String article_num, Sex sex, Long price, DressType dresstype, Integer discount) {
 		Dress dress = Dress.builder()
 				.brand(brand)
 				.article_number(article_num)
